requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.0
python-dotenv==1.1.0
google-generativeai==0.8.4
Werkzeug==3.1.3
gunicorn==21.2.0
requests==2.31.0